And another checkpoint.
