+++ java/src/Ice/src/main/java/com/zeroc/Ice/InvocationFuture.java
@@ -4,8 +4,13 @@
 
 package com.zeroc.Ice;
 
+import com.zeroc.IceInternal.CancellationHandler;
+import com.zeroc.IceInternal.Ex;
+import com.zeroc.IceInternal.Instance;
+import com.zeroc.IceInternal.RunnableThreadPoolWorkItem;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executor;
+import java.util.function.BiConsumer;
 
 /**
  * An instance of an InvocationFuture subclass is the return value of an asynchronous invocation.
@@ -19,14 +24,30 @@ public abstract class InvocationFuture<T> extends CompletableFuture<T> {
    *
    * @return True if this task is now cancelled.
    */
-  public abstract boolean cancel();
+  public boolean cancel() {
+    return cancel(false);
+  }
+
+  @Override
+  public boolean cancel(boolean mayInterruptIfRunning) {
+    //
+    // Call super.cancel(boolean) first. This sets the result of the future.
+    // Calling cancel(LocalException) also eventually attempts to complete the future
+    // (exceptionally), but this result is ignored.
+    //
+    boolean r = super.cancel(mayInterruptIfRunning);
+    cancel(new InvocationCanceledException());
+    return r;
+  }
 
   /**
    * Returns the communicator that sent the invocation.
    *
    * @return The communicator.
    */
-  public abstract Communicator getCommunicator();
+  public Communicator getCommunicator() {
+    return _communicator;
+  }
 
   /**
    * Returns the connection that was used to start the invocation, or nil if this future was not
@@ -35,7 +56,9 @@ public abstract class InvocationFuture<T> extends CompletableFuture<T> {
    *
    * @return The connection.
    */
-  public abstract Connection getConnection();
+  public Connection getConnection() {
+    return null;
+  }
 
   /**
    * Returns the proxy that was used to start the asynchronous invocation, or nil if this object was
@@ -43,17 +66,29 @@ public abstract class InvocationFuture<T> extends CompletableFuture<T> {
    *
    * @return The proxy.
    */
-  public abstract ObjectPrx getProxy();
+  public ObjectPrx getProxy() {
+    return null;
+  }
 
   /**
    * Returns the name of the operation.
    *
    * @return The operation name.
    */
-  public abstract String getOperation();
+  public final String getOperation() {
+    return _operation;
+  }
 
   /** Blocks the caller until the result of the invocation is available. */
-  public abstract void waitForCompleted();
+  public final void waitForCompleted() {
+    if (Thread.interrupted()) {
+      throw new OperationInterruptedException();
+    }
+    try {
+      join();
+    } catch (java.lang.Exception ex) {
+    }
+  }
 
   /**
    * When you start an asynchronous invocation, the Ice run time attempts to write the corresponding
@@ -64,10 +99,25 @@ public abstract class InvocationFuture<T> extends CompletableFuture<T> {
    *
    * @return True if the request has been sent, or false if the request is queued.
    */
-  public abstract boolean isSent();
+  public final boolean isSent() {
+    synchronized (this) {
+      return (_state & StateSent) > 0;
+    }
+  }
 
   /** Blocks the caller until the request has been written to the client-side transport. */
-  public abstract void waitForSent();
+  public final synchronized void waitForSent() {
+    if (Thread.interrupted()) {
+      throw new OperationInterruptedException();
+    }
+    while ((_state & StateSent) == 0 && _exception == null) {
+      try {
+        this.wait();
+      } catch (InterruptedException ex) {
+        throw new OperationInterruptedException();
+      }
+    }
+  }
 
   /**
    * Returns true if a request was written to the client-side transport without first being queued.
@@ -76,7 +126,9 @@ public abstract class InvocationFuture<T> extends CompletableFuture<T> {
    *
    * @return True if the request was sent without being queued, or false otherwise.
    */
-  public abstract boolean sentSynchronously();
+  public final boolean sentSynchronously() {
+    return _sentSynchronously; // No lock needed, immutable
+  }
 
   /**
    * Returns a future that completes when the entire request message has been accepted by the
@@ -86,8 +138,26 @@ public abstract class InvocationFuture<T> extends CompletableFuture<T> {
    * @param action Executed when the future is completed successfully or exceptionally.
    * @return A future that completes when the message has been handed off to the transport.
    */
-  public abstract CompletableFuture<Boolean> whenSent(
-      java.util.function.BiConsumer<Boolean, ? super Throwable> action);
+  public final synchronized CompletableFuture<Boolean> whenSent(
+      BiConsumer<Boolean, ? super Throwable> action) {
+    if (_sentFuture == null) {
+      _sentFuture = new CompletableFuture<>();
+    }
+
+    CompletableFuture<Boolean> r = _sentFuture.whenComplete(action);
+
+    //
+    // Check if the request has already been sent.
+    //
+    if (((_state & StateSent) > 0 || _exception != null) && !_sentFuture.isDone()) {
+      if (_exception != null) {
+        _sentFuture.completeExceptionally(_exception);
+      } else {
+        _sentFuture.complete(_sentSynchronously);
+      }
+    }
+    return r;
+  }
 
   /**
    * Returns a future that completes when the entire request message has been accepted by the
@@ -97,8 +167,10 @@ public abstract class InvocationFuture<T> extends CompletableFuture<T> {
    * @param action Executed when the future is completed successfully or exceptionally.
    * @return A future that completes when the message has been handed off to the transport.
    */
-  public abstract CompletableFuture<Boolean> whenSentAsync(
-      java.util.function.BiConsumer<Boolean, ? super Throwable> action);
+  public final synchronized CompletableFuture<Boolean> whenSentAsync(
+      BiConsumer<Boolean, ? super Throwable> action) {
+    return whenSentAsync(action, null);
+  }
 
   /**
    * Returns a future that completes when the entire request message has been accepted by the
@@ -109,6 +181,295 @@ public abstract class InvocationFuture<T> extends CompletableFuture<T> {
    * @param executor The executor to use for asynchronous execution.
    * @return A future that completes when the message has been handed off to the transport.
    */
-  public abstract CompletableFuture<Boolean> whenSentAsync(
-      java.util.function.BiConsumer<Boolean, ? super Throwable> action, Executor executor);
+  public final synchronized CompletableFuture<Boolean> whenSentAsync(
+      BiConsumer<Boolean, ? super Throwable> action, Executor executor) {
+    if (_sentFuture == null) {
+      _sentFuture = new CompletableFuture<>();
+    }
+
+    CompletableFuture<Boolean> r;
+    if (executor == null) {
+      r = _sentFuture.whenCompleteAsync(action);
+    } else {
+      r = _sentFuture.whenCompleteAsync(action, executor);
+    }
+
+    //
+    // Check if the request has already been sent.
+    //
+    if (((_state & StateSent) > 0 || _exception != null) && !_sentFuture.isDone()) {
+      if (_exception != null) {
+        _sentFuture.completeExceptionally(_exception);
+      } else {
+        _sentFuture.complete(_sentSynchronously);
+      }
+    }
+    return r;
+  }
+
+  public final void invokeSent() {
+    try {
+      synchronized (this) {
+        if (_sentFuture != null && !_sentFuture.isDone()) {
+          _sentFuture.complete(_sentSynchronously);
+        }
+      }
+
+      if (_doneInSent) {
+        markCompleted();
+      }
+    } catch (java.lang.RuntimeException ex) {
+      warning(ex);
+    } catch (java.lang.Error exc) {
+      error(exc);
+      if (!(exc instanceof java.lang.AssertionError || exc instanceof java.lang.OutOfMemoryError)) {
+        throw exc;
+      }
+    }
+
+    if (_observer != null) {
+      ObjectPrx proxy = getProxy();
+      if (proxy == null || !proxy.ice_isTwoway()) {
+        _observer.detach();
+        _observer = null;
+      }
+    }
+  }
+
+  protected abstract void markCompleted();
+
+  public final void invokeCompleted() {
+    try {
+      if (_exception != null) {
+        synchronized (this) {
+          if (_sentFuture != null && !_sentFuture.isDone()) {
+            _sentFuture.completeExceptionally(_exception);
+          }
+        }
+        completeExceptionally(_exception);
+      } else {
+        markCompleted();
+      }
+    } catch (RuntimeException ex) {
+      warning(ex);
+    } catch (AssertionError exc) {
+      error(exc);
+    } catch (OutOfMemoryError exc) {
+      error(exc);
+    }
+
+    if (_observer != null) {
+      _observer.detach();
+      _observer = null;
+    }
+  }
+
+  public final void invokeCompletedAsync() {
+    //
+    // CommunicatorDestroyedException is the only exception that can propagate directly from this
+    // method.
+    //
+    _instance
+        .clientThreadPool()
+        .dispatch(
+            new RunnableThreadPoolWorkItem(_cachedConnection) {
+              @Override
+              public void run() {
+                invokeCompleted();
+              }
+            });
+  }
+
+  public synchronized void cancelable(final CancellationHandler handler) {
+    if (_cancellationException != null) {
+      try {
+        throw _cancellationException;
+      } finally {
+        _cancellationException = null;
+      }
+    }
+    _cancellationHandler = handler;
+  }
+
+  protected InvocationFuture(Communicator communicator, Instance instance, String op) {
+    _communicator = communicator;
+    _instance = instance;
+    _operation = op;
+    _state = 0;
+    _sentSynchronously = false;
+    _doneInSent = false;
+    _synchronous = false;
+    _exception = null;
+  }
+
+  protected void cacheMessageBuffers() {}
+
+  protected boolean sent(boolean done) {
+    synchronized (this) {
+      assert (_exception == null);
+
+      boolean alreadySent = (_state & StateSent) != 0;
+      _state |= StateSent;
+      if (done) {
+        _state |= StateDone | StateOK;
+        _cancellationHandler = null;
+        _doneInSent = true;
+
+        //
+        // For oneway requests after the data has been sent
+        // the buffers can be reused unless this is a
+        // collocated invocation. For collocated invocations
+        // the buffer won't be reused because it has already
+        // been marked as cached in invokeCollocated.
+        //
+        cacheMessageBuffers();
+      }
+
+      if (_synchronous && done) {
+        if (_observer != null) {
+          _observer.detach();
+          _observer = null;
+        }
+
+        markCompleted();
+        return false;
+      } else {
+        this.notifyAll();
+        boolean invoke = (!alreadySent && _sentFuture != null || done) && !_synchronous;
+        return invoke;
+      }
+    }
+  }
+
+  protected boolean finished(boolean ok, boolean invoke) {
+    synchronized (this) {
+      _state |= StateDone;
+      if (ok) {
+        _state |= StateOK;
+      }
+      _cancellationHandler = null;
+
+      invoke &= !_synchronous;
+      if (!invoke && _observer != null) {
+        _observer.detach();
+        _observer = null;
+      }
+
+      if (!invoke) {
+        if (_exception != null) {
+          completeExceptionally(_exception);
+        } else {
+          markCompleted();
+        }
+        return false;
+      } else {
+        this.notifyAll();
+        return invoke;
+      }
+    }
+  }
+
+  protected boolean finished(com.zeroc.Ice.Exception ex) {
+    synchronized (this) {
+      _state |= StateDone;
+      _exception = ex;
+      _cancellationHandler = null;
+      if (_observer != null) {
+        _observer.failed(ex.ice_id());
+      }
+
+      boolean invoke = !_synchronous;
+      if (!invoke && _observer != null) {
+        _observer.detach();
+        _observer = null;
+      }
+
+      if (!invoke) {
+        if (_exception != null) {
+          completeExceptionally(_exception);
+        } else {
+          markCompleted();
+        }
+        return false;
+      } else {
+        this.notifyAll();
+        return invoke;
+      }
+    }
+  }
+
+  public final void invokeSentAsync() {
+    //
+    // This is called when it's not safe to call the sent callback
+    // synchronously from this thread. Instead the future is completed
+    // asynchronously from a client in the client thread pool.
+    //
+    dispatch(() -> invokeSent());
+  }
+
+  protected void cancel(LocalException ex) {
+    CancellationHandler handler;
+    synchronized (this) {
+      if (_cancellationHandler == null) {
+        _cancellationException = ex;
+        return;
+      }
+      handler = _cancellationHandler;
+    }
+    handler.asyncRequestCanceled((com.zeroc.IceInternal.OutgoingAsyncBase) this, ex);
+  }
+
+  protected com.zeroc.Ice.Instrumentation.InvocationObserver getObserver() {
+    return _observer;
+  }
+
+  protected void dispatch(final Runnable runnable) {
+    try {
+      _instance
+          .clientThreadPool()
+          .dispatch(
+              new RunnableThreadPoolWorkItem(_cachedConnection) {
+                @Override
+                public void run() {
+                  runnable.run();
+                }
+              });
+    } catch (CommunicatorDestroyedException ex) {
+    }
+  }
+
+  private void warning(RuntimeException ex) {
+    if (_instance.initializationData().properties.getIcePropertyAsInt("Ice.Warn.AMICallback") > 0) {
+      String s = "exception raised by AMI callback:\n" + Ex.toString(ex);
+      _instance.initializationData().logger.warning(s);
+    }
+  }
+
+  private void error(Error error) {
+    String s = "error raised by AMI callback:\n" + Ex.toString(error);
+    _instance.initializationData().logger.error(s);
+  }
+
+  protected final Instance _instance;
+  protected com.zeroc.Ice.Instrumentation.InvocationObserver _observer;
+  protected Connection _cachedConnection;
+  protected boolean _sentSynchronously;
+  protected boolean _doneInSent;
+  // True if this AMI request is being used for a generated synchronous invocation.
+  protected boolean _synchronous;
+  protected CompletableFuture<Boolean> _sentFuture;
+
+  protected final Communicator _communicator;
+  protected final String _operation;
+
+  protected com.zeroc.Ice.Exception _exception;
+
+  private CancellationHandler _cancellationHandler;
+  private LocalException _cancellationException;
+
+  protected static final byte StateOK = 0x1;
+  protected static final byte StateDone = 0x2;
+  protected static final byte StateSent = 0x4;
+  protected static final byte StateCachedBuffers = 0x08;
+  protected byte _state;
 }
